@Girgias 2023-04-11
Technically, any comparison involving objects can throw, see: https://3v4l.org/64TZU