@Girgias 2023-04-11
> Thanks @Girgias for taking a look. Tim made me aware of the error behaviour just recently. That is why there is a check that checks whether there is a user error handler to disallow the "input=output" optimization for array_unique.
> 
> In fact there are two optimizations
> 
>     * RC1 optimization: if there is only one reference to an array, it's because it was passed as an argument and therefore it can be replaced. Even if there is an exception, a partial or in-place modification cannot be observed because the assignment never happened and the return value is destroyed.
> 
>     * input=output optimization: if the input gets overwritten by the output, it is safe to do so if there are no exception handlers or other user code that can run in between the function call and the assignment.
> 
> 
> The four functions: array_merge, array_unique, array_replace, and array_intersect all perform the RC1 optimization with this patch.
> 
> array_unique only performs the input=output optimization if the sorting algorithm is SORT_REGULAR or SORT_NUMERIC _and_ no user error handler is installed. array_merge and array_replace can do the input=output optimization safely in the non-recursive case.
> 
> Do you think there is something I'm overlooking?

So any internal object can define its own comparison handler, and thus can throw. If I modify the previous example to:
```php
$o = gmp_init(10);
$fruits = array($o, []);
sort($fruits);
```
The call to ``sort()`` will throw a TypeError:
```
Fatal error: Uncaught TypeError: Number must be of type GMP|string|int, array given in /tmp/preview:5
```

So I'm not sure you can even safely assume that comparisons are not going to throw an exception even for SORT_NUMERIC and SORT_REGULAR, except if no objects are present, but that means traversing the array :/ (or us having typed arrays)